Refactored from using navigation events to display screens to ScreenDisplayCommands
Some thoughts:
I figured that raising events is more of a "something has happened" and then other things can *react* to the event.
However, I felt that displaying a screen was more of a _proactive_ action than a _reactive_ action.
So I switched it about a bit and implemented a sort of Command Pattern based approach to displaying the screens.
This allowed the removal of the NavigationBus (however, I'll probably reintroduce an event bus for things like "ScreenDidDisplay" that the Stats system can react to).

Problems:
- The fragment boundary problem persists... how do I get my ScreenManager into the first fragment and allow it to be used in a ScreenDisplayCommand that shows the next screen? I've implemented a quick and dirty ServiceLocator for the moment but I'm looking for ways to improve this.